refactor(format): one byte count, whatever integer the caller is holding
format.Bytes took a uint64, and nothing that measures bytes in Go holds one:
io.Copy returns int64, FileInfo.Size is int64, len is int, and a server's own
stat is whatever its client library declared. So every caller wrote the
widening itself — twenty-one sites here, thirty-two in the plugins — and
because gosec reads int64 -> uint64 as an overflow nothing has ruled out, five
of them carried a //nolint repeating one sentence about one fact ("a file size
is never negative", "a directory's size, never negative", "a sum of file sizes
is never negative") while three more clamped with max(n, 0) in case that
sentence was wrong.

Widening a byte count is the function's business now. Five suppressions and a
clamp go with it, and no new suppression arrives in their place: gosec does not
flag the conversion inside the generic, checked by stripping the directives and
re-running both lint passes rather than by leaving them in to be safe.

A negative count keeps its sign instead of wrapping into the unsigned range.
Under the old signature -1 rendered as "16.0 EiB", a figure plausible enough to
act on; it reads "-1 B" now. The minimum of every signed type survives the
negation, which is what -(n+1)+1 is for and what the int8 case pins.

Measured, not assumed: the stripped linux/amd64 binary is 37,810,336 bytes
before and after — different binaries, identical size — because the integer
instantiations share one gcshape and the arithmetic stays in a single
non-generic function.

The plugins' thirty-two casts, and the G115 exclusion their lint config carries
for them, come out once they pin a release with this in it.

+++ builtin/kv/history.go
@@ -102,10 +102,10 @@ func runHistory(_ context.Context, req plugin.Request) (view.View, error) {
 		{Name: "Replaced", Kind: view.KindDuration},
 		{Name: "Source"},
 	}}
-	t.Rows = append(t.Rows, []string{head, e.Kind, format.Bytes(uint64(len(e.Value))),
+	t.Rows = append(t.Rows, []string{head, e.Kind, format.Bytes(len(e.Value)),
 		e.Description, itemstore.Age(e.Updated), "", e.origin()})
 	for i, r := range e.Previous {
-		t.Rows = append(t.Rows, []string{strconv.Itoa(i + 1), r.Kind, format.Bytes(uint64(len(r.Value))),
+		t.Rows = append(t.Rows, []string{strconv.Itoa(i + 1), r.Kind, format.Bytes(len(r.Value)),
 			r.Description, itemstore.Age(r.Updated), itemstore.Age(r.Retired), r.Origin})
 	}
 	t.Total = len(t.Rows)
@@ -224,7 +224,7 @@ func removedTable(s store) view.View {
 	}}
 	for _, k := range names {
 		r := s.Removed[k]
-		t.Rows = append(t.Rows, []string{k, r.Kind, format.Bytes(uint64(len(r.Value))), r.Description,
+		t.Rows = append(t.Rows, []string{k, r.Kind, format.Bytes(len(r.Value)), r.Description,
 			itemstore.Age(r.RemovedAt)})
 	}
 	t.Total = len(t.Rows)
